+++ packages/sdk/src/sdk-blockchains/ethereum/common/index.ts
@@ -18,7 +18,8 @@ import { toBn } from "@rarible/utils/build/bn"
 import type { AssetType as EthereumAssetType, Part } from "@rarible/ethereum-api-client"
 import type { Order } from "@rarible/ethereum-api-client/build/models"
 import type { EthereumTransaction } from "@rarible/ethereum-provider"
-import type { CurrencyType, RequestCurrency } from "../../../common/domain"
+import type { CurrencyType } from "../../../common/domain"
+import type { RequestCurrencyAssetType } from "../../../common/domain"
 
 export type EVMBlockchain = Blockchain.ETHEREUM | Blockchain.POLYGON
 export const EVMBlockchains: EVMBlockchain[] = [
@@ -28,7 +29,7 @@ export const EVMBlockchains: EVMBlockchain[] = [
 
 export type CreateEthereumCollectionResponse = { tx: EthereumTransaction, address: Address }
 
-export function getEthTakeAssetType(currency: RequestCurrency) {
+export function getEthTakeAssetType(currency: RequestCurrencyAssetType) {
 	switch (currency["@type"]) {
 		case "ERC20":
 			return {