+++ packages/sdk/src/sdk-blockchains/tezos/balance.ts
@@ -1,21 +1,28 @@
-import type { TezosProvider } from "@rarible/tezos-sdk"
+import type { TezosProvider, TezosNetwork } from "@rarible/tezos-sdk"
 import type { UnionAddress } from "@rarible/types"
-import type { AssetType } from "@rarible/api-client"
 import type { BigNumberValue } from "@rarible/utils"
 // eslint-disable-next-line camelcase
-import { get_balance } from "@rarible/tezos-sdk"
+import { get_balance, unwrap, wrap } from "@rarible/tezos-sdk"
 import BigNumber from "bignumber.js"
+import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
+import { BlockchainTezosTransaction } from "@rarible/sdk-transaction"
+import type { ConvertRequest } from "../../types/balances"
+import type { RequestCurrency } from "../../common/domain"
+import { getCurrencyAssetType } from "../../common/get-currency-asset-type"
 import type { MaybeProvider } from "./common"
-import { getTezosAddress, getTezosAssetType } from "./common"
+import { getRequiredProvider, getTezosAddress, getTezosAssetType } from "./common"
 
 export class TezosBalance {
 	constructor(
 		private provider: MaybeProvider<TezosProvider>,
+		private network: TezosNetwork,
 	) {
 		this.getBalance = this.getBalance.bind(this)
+		this.convert = this.convert.bind(this)
 	}
 
-	async getBalance(address: UnionAddress, assetType: AssetType): Promise<BigNumberValue> {
+	async getBalance(address: UnionAddress, currency: RequestCurrency): Promise<BigNumberValue> {
+		const assetType = getCurrencyAssetType(currency)
 		const tezosAssetType = getTezosAssetType(assetType)
 		if (tezosAssetType.asset_class !== "XTZ" && tezosAssetType.asset_class !== "FT") {
 			throw new Error("Unsupported asset type")
@@ -30,5 +37,16 @@ export class TezosBalance {
 				tezosAssetType
 			)
 		)
+	}
+
+	async convert(request: ConvertRequest): Promise<IBlockchainTransaction> {
+		const provider = getRequiredProvider(this.provider)
+		if (request.isWrap) {
+			const tx = await wrap(provider, new BigNumber(request.value))
+			return new BlockchainTezosTransaction(tx, this.network)
+		} else {
+			const tx = await unwrap(provider, new BigNumber(request.value))
+			return new BlockchainTezosTransaction(tx, this.network)
+		}
 	}
 }