+++ scripts/build.sh
@@ -12,3 +12,4 @@ yarn run build-connector-fortmatic
 yarn run build-connector-torus
 yarn run build-connector-walletlink
 yarn run build-connector-walletconnect
+yarn run build-connector-phantom